Issue 14 : Log History dialog shows last column "Description" too small.
